Move some methods up to superclass.
